My E-Business2 : Stock quantity in database

+++ FULL STACK/Exos/MyEbusiness (plein)/src/Controller/AdminCategoryController.php
@@ -8,6 +8,8 @@ use App\Form\SearchType;
 use App\Form\CategoryType;
 use App\Repository\ProductRepository;
 use App\Repository\CategoryRepository;
+use App\Repository\OrderDetailsRepository;
+use App\Service\Cart\CartService;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -17,8 +19,13 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 class AdminCategoryController extends AbstractController
 {
     #[Route('/', name: 'app_admin_category_index', methods: ['GET'])]
-    public function index(CategoryRepository $categoryRepository, ProductRepository $productRepository): Response
+    public function index(CategoryRepository $categoryRepository, ProductRepository $productRepository, OrderDetailsRepository $orderDetails, CartService $cartService): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            $this->addFlash('error', 'Access denied');
+            return $this->redirectToRoute('login');  
+        }
+
         $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
 
         $categories = $categoryRepository->findAll();
@@ -29,6 +36,9 @@ class AdminCategoryController extends AbstractController
         $discount2 =$productRepository->findBy(['discount' => true]);
         
         return $this->render('admin_category/index.html.twig', [
+            'items'     => $cartService->getFullCart($orderDetails),
+            'count'     => $cartService->getItemCount($orderDetails),
+            'total' => $cartService->getTotal($orderDetails),
             'products' => $products,
             'products2' => $products2,
             'categories' => $categories,
@@ -38,9 +48,14 @@ class AdminCategoryController extends AbstractController
     }
 
     #[Route('/new', name: 'app_admin_category_new', methods: ['GET', 'POST'])]
-    public function new(CategoryRepository $categoryRepository, Request $request,ProductRepository $productRepository): Response
+    public function new(CategoryRepository $categoryRepository, Request $request,ProductRepository $productRepository, OrderDetailsRepository $orderDetails, CartService $cartService): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            $this->addFlash('error', 'Access denied');
+            return $this->redirectToRoute('login');  
+        }
         $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
+
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
@@ -59,6 +74,9 @@ class AdminCategoryController extends AbstractController
         }
 
         return $this->renderForm('admin_category/new.html.twig', [
+            'items'     => $cartService->getFullCart($orderDetails),
+            'count'     => $cartService->getItemCount($orderDetails),
+            'total' => $cartService->getTotal($orderDetails),
             'category' => $category,
             'form' => $form,
             'products' => $products,
@@ -70,8 +88,12 @@ class AdminCategoryController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_admin_category_show', methods: ['GET'])]
-    public function show(Category $category, CategoryRepository $categoryRepository,ProductRepository $productRepository): Response
+    public function show(Category $category, CategoryRepository $categoryRepository,ProductRepository $productRepository, OrderDetailsRepository $orderDetails, CartService $cartService): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            $this->addFlash('error', 'Access denied');
+            return $this->redirectToRoute('login');  
+        }
         $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
 
         $categories = $categoryRepository->findAll();
@@ -82,6 +104,9 @@ class AdminCategoryController extends AbstractController
         $discount2 =$productRepository->findBy(['discount' => true]);
 
         return $this->render('admin_category/show.html.twig', [
+            'items'     => $cartService->getFullCart($orderDetails),
+            'count'     => $cartService->getItemCount($orderDetails),
+            'total' => $cartService->getTotal($orderDetails),
             'category' => $category,
             'products' => $products,
             'products2' => $products2,
@@ -92,8 +117,12 @@ class AdminCategoryController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_admin_category_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Category $category, CategoryRepository $categoryRepository,ProductRepository $productRepository): Response
+    public function edit(Request $request, Category $category, CategoryRepository $categoryRepository,ProductRepository $productRepository, OrderDetailsRepository $orderDetails, CartService $cartService): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            $this->addFlash('error', 'Access denied');
+            return $this->redirectToRoute('login');  
+        }
         $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
 
         $form = $this->createForm(CategoryType::class, $category);
@@ -113,6 +142,9 @@ class AdminCategoryController extends AbstractController
         }
 
         return $this->renderForm('admin_category/edit.html.twig', [
+            'items'     => $cartService->getFullCart($orderDetails),
+            'count'     => $cartService->getItemCount($orderDetails),
+            'total' => $cartService->getTotal($orderDetails),
             'category' => $category,
             'form' => $form,
             'products' => $products,
@@ -126,7 +158,12 @@ class AdminCategoryController extends AbstractController
     #[Route('/{id}', name: 'app_admin_category_delete', methods: ['POST'])]
     public function delete(Request $request, Category $category, CategoryRepository $categoryRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            $this->addFlash('error', 'Access denied');
+            return $this->redirectToRoute('login');  
+        }
         $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
+        
         if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
             $categoryRepository->remove($category, true);
         }
